add WIP

+++ police_reports_berlin/pipelines.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from datetime import datetime
+
 # Define your item pipelines here
 #
 # Don't forget to add your pipeline to the ITEM_PIPELINES setting
@@ -11,11 +13,12 @@ class MongoPipeline(object):
     client = None
     db = None
 
-    def __init__(self, mongo_enabled, mongo_uri, mongo_db, mongo_collection):
+    def __init__(self, mongo_enabled, mongo_uri, mongo_db, mongo_collection, mongo_drop_collection):
         self.mongo_enabled = mongo_enabled
         self.mongo_uri = mongo_uri
         self.mongo_db = mongo_db
         self.mongo_collection = mongo_collection
+        self.mongo_drop_collection = mongo_drop_collection
 
     @classmethod
     def from_crawler(cls, crawler):
@@ -23,14 +26,18 @@ class MongoPipeline(object):
             mongo_enabled=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_ENABLED'),
             mongo_uri=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_URI'),
             mongo_db=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_DATABASE'),
-            mongo_collection=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_COLLECTION')
+            mongo_collection=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_COLLECTION'),
+            mongo_drop_collection=crawler.settings.get('SCRAPY_POLICE_REPORTS_CRAWLER_MONGO_DROP_COLLECTION')
         )
 
     def open_spider(self, spider):
         if self.mongo_enabled:
             self.client = pymongo.MongoClient(self.mongo_uri)
             self.db = self.client[self.mongo_db]
-            self.db[self.mongo_collection].drop()
+
+            if self.mongo_drop_collection:
+                self.db[self.mongo_collection].drop()
+                print('Dropped collection')
 
     def close_spider(self, spider):
         if self.mongo_enabled:
@@ -38,8 +45,18 @@ class MongoPipeline(object):
 
     def process_item(self, item, spider):
         if self.mongo_enabled:
-            # Only crawl if url was not crawled yet
-            if not self.db[self.mongo_collection].find_one({'url': item['url']}):
+            item['updated'] = datetime.now()
+            doc = self.db[self.mongo_collection].find_one({'url': item['url']})
+
+            if not doc:
+                item['created'] = datetime.now()
                 self.db[self.mongo_collection].insert_one(dict(item))
-            
-        return item
+
+                print('Inserted new document')
+            else:
+                # self.db[self.mongo_drop_collection].update_one(
+                #    {'_id': doc['_id']},
+                #    {'$set': dict(item)}
+                # )
+
+                print('Updated existing document')
